move utils.span to tests
