+++ tests/contrib/django/conftest.py
@@ -6,8 +6,7 @@ import pytest
 from ddtrace import Pin
 from ddtrace.contrib.django import patch
 
-from ...utils.span import TracerSpanContainer
-from ... import DummyTracer
+from ... import DummyTracer, TracerSpanContainer
 
 # We manually designate which settings we will be using in an environment variable
 # This is similar to what occurs in the `manage.py`
